+++ dicom_validator/spec_reader/condition_parser.py
@@ -1,10 +1,11 @@
 import re
 from collections import OrderedDict
+from typing import Dict, Tuple, Optional, List
 
 from dicom_validator.spec_reader.condition import Condition
 
 
-class ConditionParser(object):
+class ConditionParser:
     """Parses the description of type C modules and type 1C and 2C attributes.
     Creates a Condition object representing the condition(s) in the
     description provided that:
@@ -23,6 +24,7 @@ class ConditionParser(object):
     '<' - required if the given tag value is less than the given value
     '=>' - required if the given tag points to one of the given tags
     """
+    all_conditions = {}
 
     tag_expression = re.compile(
         r'(the value of )?(?P<name>[a-zA-Z1-9 \'\-]+)'
@@ -65,23 +67,28 @@ class ConditionParser(object):
         ('or', 'or')
     ])
 
-    def __init__(self, dict_info):
+    def __init__(self, dict_info: Dict) -> None:
         self._dict_info = dict_info
 
-    def parse(self, condition):
+    def parse(self, condition_str: str) -> Condition:
         """Parse the given condition string and return a Condition object
          with the required attributes.
         """
         condition_prefixes = ('required if ', 'shall be present if ')
         for prefix in condition_prefixes:
-            index = condition.lower().find(prefix)
+            index = condition_str.lower().find(prefix)
             if index >= 0:
-                condition = condition[len(prefix) + index:]
-                condition = self._fix_condition(condition)
-                return self._parse_tag_expressions(condition)
+                condition_str = condition_str[len(prefix) + index:]
+                condition_str = self._fix_condition(condition_str)
+                condition = self._parse_tag_expressions(condition_str)
+                self.__class__.all_conditions[condition_str] = \
+                    condition.to_string(self._dict_info)
+                return condition
+        self.__class__.all_conditions[condition_str] = Condition(ctype='U')
         return Condition(ctype='U')
 
-    def _parse_tag_expression(self, condition):
+    def _parse_tag_expression(
+            self, condition: str) -> Tuple[Condition, Optional[str]]:
         operator_text = None
         op_offset = None
         for operator in self.operators:
@@ -89,20 +96,20 @@ class ConditionParser(object):
             if offset > 0 and (op_offset is None or offset < op_offset):
                 op_offset = offset
                 operator_text = operator
-        if operator_text is None:
+        if operator_text is None or op_offset is None:
             return Condition(ctype='U'), None
         operator = self.operators[operator_text]
         rest = condition[op_offset + len(operator_text):]
-        if self.operators[operator_text] in ('=', '!=', '>', '<'):
+        if operator in ('=', '!=', '>', '<'):
             values, rest = self._parse_tag_values(rest)
-        elif self.operators[operator_text] == '=>':
-            value_string, rest = self.extract_value_string(rest)
+        elif operator == '=>':
+            value_string, rest = self._split_value_part(rest)
             tag, _ = self._parse_tag(value_string)
             if tag is None:
                 return Condition(ctype='U'), None
             values, rest = [str(self._tag_id(tag))], rest
         else:
-            values, rest = None, rest.strip()
+            values, rest = [], rest.strip()
         result = self._parse_tags(condition[:op_offset], operator, values)
         if not result:
             return Condition(ctype='U'), None
@@ -111,7 +118,8 @@ class ConditionParser(object):
                                in condition[op_offset:].lower() else 'MN')
         return result, rest
 
-    def _get_other_condition(self, condition_string):
+    def _get_other_condition(
+            self, condition_string: str) -> Optional[Condition]:
         other_cond_texts = [
             'may be present otherwise if ', 'may be present if '
         ]
@@ -120,13 +128,15 @@ class ConditionParser(object):
             if index >= 0:
                 return self._parse_tag_expressions(
                     condition_string[index + len(condition_marker):])
+        return None
 
     @staticmethod
-    def _tag_id(tag_id_string):
+    def _tag_id(tag_id_string: str) -> int:
         group, element = tag_id_string[1:-1].split(',')
         return (int(group, 16) << 16) + int(element, 16)
 
-    def _parse_tag(self, tag_string):
+    def _parse_tag(
+            self, tag_string: str) -> Tuple[Optional[str], Optional[int]]:
         match = self.tag_expression.match(tag_string.strip())
         if match:
             value_index = (0 if match.group('index') is None
@@ -139,60 +149,98 @@ class ConditionParser(object):
                     return tag_id, value_index
         return None, None
 
-    def _parse_tag_values(self, value_string):
-        value_string, rest = self.extract_value_string(value_string)
-        values = value_string.split(', ')
-        tag_values = []
-        for value in values:
-            if ' or ' in value:
-                tag_values.extend(value.split(' or '))
-            elif value.startswith('or '):
-                tag_values.append(value[3:])
-            else:
-                tag_values.append(value)
+    def _parse_tag_values(
+            self, value_string: str) -> Tuple[List[str], str]:
+        value_part, rest = self._split_value_part(value_string)
         values = []
-        for value in tag_values:
-            value = value.strip()
-            if value.startswith('"') and value.endswith('"'):
-                value = value[1:-1].strip()
-            values.append(value)
+        value_index = 0
+        last_or = None
+        while value_index >= 0:
+            value_index = -1
+            current_or = None
+            for or_phrase in self._valid_or_expressions_after_or_expression(
+                last_or
+            ):
+                index = value_part.find(or_phrase)
+                if index >= 0 and (value_index < 0 or index < value_index):
+                    value_index = index
+                    next_index = index + len(or_phrase)
+                    current_or = or_phrase
+            if value_index > 0:
+                value = self._get_const_value(value_part[:value_index],
+                                             value_part[value_index:] + rest)
+                if value is None:
+                    return values, current_or + value_part + rest
+                values.append(value)
+                value_part = value_part[next_index:]
+                last_or = current_or
+        value = self._get_const_value(value_part, rest)
+        if value is None:
+            return values, value_part + rest
+        values.append(value)
         return values, rest
 
-    def extract_value_string(self, value_string):
-        # remove stuff that breaks parser
-        value_string = value_string.replace('(Legacy Converted)', '')
-        start_index = 0
-        rest = None
-        while True:
-            end_index = -1
-            # todo: handle or
-            for end_char in (';', '.', 'and ', ', or ', ' or '):
-                char_index = value_string.find(end_char, start_index)
-                if end_index < 0 or 0 <= char_index < end_index:
-                    end_index = char_index
-            apo_index = value_string.find('"', start_index)
-            if end_index < 0:
-                break
-            if 0 <= apo_index < end_index:
-                start_index = value_string.find('"', apo_index + 1) + 1
-                continue
-            if end_index > 0:
-                if value_string.find(' or ') in [end_index, end_index + 1]:
-                    # differentiate between several values and several
-                    # conditions - check if the rest is a condition
-                    or_cond = self._parse_tag_expressions(
-                        value_string[end_index + 3:])
-                    if or_cond.type == 'U':
-                        start_index = end_index + 4
-                        continue
-                rest = value_string[end_index:].strip()
-                value_string = value_string[:end_index]
-                break
-        return value_string, rest
+    def _split_value_part(self, value_string: str) -> Tuple[str, str]:
+        value_string = value_string.strip()
+        end_index = self._end_index_for_stop_chars(
+            value_string, [';', '.', ', and ', ' and '])
+        return value_string[:end_index], value_string[end_index:]
 
-    def _parse_tag_expressions(self, condition, nested=False):
+    @staticmethod
+    def _valid_or_expressions_after_or_expression(
+            previous_expr: Optional[str]) -> List[str]:
+        if previous_expr is None:
+            return [', ', ' or ']
+        if previous_expr == ' or ':
+            return [' or ']
+        if previous_expr == ', ':
+            return [', or ', ', ', ' or ']
+        return []
+
+    @staticmethod
+    def _index_is_inside_string(value: str, index: int) -> bool:
+        in_string = False
+        apo_index = 0
+        while apo_index >= 0:
+            apo_index = value.find('"', apo_index, index)
+            if apo_index >= 0:
+                apo_index += 1
+                in_string = not in_string
+        return in_string
+
+    def _get_const_value(self, value: str, rest: str) -> str:
+        value = value.strip()
+        if value[0] == value[-1] == '"':
+            return value[1:-1]
+        if re.match('^[A-Z0-9_ ]+$', value) is not None:
+            return value
+        tag, index = self._parse_tag(value)
+        if tag is not None:
+            return value
+
+    def _end_index_for_stop_chars(
+            self, value: str, stop_chars: List[str]) -> int:
+        end_index = len(value)
+        for stop_char in stop_chars:
+            start_index = 0
+            stop_index = 0
+            while stop_index >= 0:
+                stop_index = value.find(
+                    stop_char, start_index, end_index)
+                if stop_index < 0:
+                    break
+                if self._index_is_inside_string(value, stop_index):
+                    start_index = stop_index + 1
+                else:
+                    end_index = stop_index
+                    break
+        return end_index
+
+    def _parse_tag_expressions(
+            self, condition: str, nested: bool = False) -> Condition:
         result, rest = self._parse_tag_expression(condition)
         if rest is not None:
+            rest = rest.strip()
             if rest.startswith(', '):
                 rest = rest[2:]
             logical_op = None
@@ -219,7 +267,9 @@ class ConditionParser(object):
                 result.other_condition = other_cond
         return result
 
-    def _parse_tags(self, condition, operator, values):
+    def _parse_tags(
+            self, condition: str, operator: str, values: List[str]
+    ) -> Optional[Condition]:
         # this handles only a few cases that are actually found
         if ', and ' in condition:
             return self._parse_tag_composition(
@@ -235,21 +285,29 @@ class ConditionParser(object):
                 condition, operator, values, 'or')
         return self._result_from_tag_string(condition, operator, values)
 
-    def _parse_tag_composition(self, condition, operator, values, logical_op):
+    def _parse_tag_composition(
+            self, condition_str: str, operator: str,
+            values: List[str], logical_op: str
+    ) -> Optional[Condition]:
         split_string = ', {} '.format(logical_op)
-        conditions = condition.split(split_string)
+        conditions = condition_str.split(split_string)
         result0 = self._parse_tags(conditions[0], operator, values)
-        if not result0:
-            result = self._parse_tags(condition.replace(
+        if result0 is None:
+            result = self._parse_tags(condition_str.replace(
                 split_string, split_string.replace(',', '')), operator, values)
         else:
             result = Condition()
             cond_list = self._condition_list(logical_op, result)
             cond_list.append(result0)
-            cond_list.append(self._parse_tags(conditions[1], operator, values))
+            condition = self._parse_tags(conditions[1], operator, values)
+            if condition is not None:
+                cond_list.append(condition)
         return result
 
-    def _parse_multiple_tags(self, condition, operator, values, logical_op):
+    def _parse_multiple_tags(
+            self, condition: str, operator: str,
+            values: List[str], logical_op: str
+    ) -> Optional[Condition]:
         condition = condition.replace(' {} '.format(logical_op), ', ')
         result = Condition()
         cond_list = self._condition_list(logical_op, result)
@@ -260,23 +318,28 @@ class ConditionParser(object):
                 cond_list.append(tag_result)
         if len(cond_list) > 1:
             return result
+        return None
 
     @staticmethod
-    def _condition_list(logical_op, result):
+    def _condition_list(logical_op: str, result: Condition) -> List[Condition]:
         cond_list = (result.and_conditions if logical_op == 'and'
                      else result.or_conditions)
         return cond_list
 
-    def _result_from_tag_string(self, tag_string, operator, values):
+    def _result_from_tag_string(
+            self, tag_string: str, operator: str, values: List[str]
+    ) -> Optional[Condition]:
         tag, index = self._parse_tag(tag_string)
         if tag is not None:
             result = Condition(tag=tag, index=index, operator=operator)
             if values:
                 result.values = values
             return result
+        return None
 
     @staticmethod
-    def _fix_condition(condition):
+    def _fix_condition(condition: str) -> str:
+        condition = condition.replace('(Legacy Converted)', '')
         index = condition.lower().find(' may be present otherwise')
         if index > 0:
             if condition[index - 1] == ',':